requirements.txt
tensorflow==2.16.1
Flask==3.0.3
pandas==2.2.2
torch==2.4.0
keras==3.3.3
opencv-python==4.10.0.84
scikit-learn==1.2.2
pillow==10.3.0